Update firmware to version 4.2.1

+++ firmware/src/micropython/drivers/cyw43/cyw43_lwip.c
@@ -24,10 +24,12 @@
  * THE SOFTWARE.
  */
 
+#include <stdio.h>
 #include <string.h>
+#include <stdio.h>
 
 #include "py/mphal.h"
-#include "lib/netutils/netutils.h"
+#include "shared/netutils/netutils.h"
 #include "lwip/etharp.h"
 #include "lwip/dns.h"
 #include "lwip/apps/mdns.h"
@@ -49,7 +51,7 @@ STATIC void cyw43_ethernet_trace(cyw43_t *self, struct netif *netif, size_t len,
         }
 
         if (self->trace_flags & CYW43_TRACE_MAC) {
-            printf("[% 8d] ETH%cX itf=%c%c len=%u", mp_hal_ticks_ms(), is_tx ? 'T' : 'R', netif->name[0], netif->name[1], len);
+            printf("[% 8d] ETH%cX itf=%c%c len=%u", (int)mp_hal_ticks_ms(), is_tx ? 'T' : 'R', netif->name[0], netif->name[1], len);
             printf(" MAC type=%d subtype=%d data=", buf[0] >> 2 & 3, buf[0] >> 4);
             for (size_t i = 0; i < len; ++i) {
                 printf(" %02x", buf[i]);
